fix(hints): add 2>/dev/null prohibition to all caller hints (#1622)
Callers habitually pipe csa stderr to /dev/null to suppress WARNING
messages, but this also hides ERROR messages from invalid arguments
(e.g., --timeout on session wait), causing silent retry loops that
waste thousands of tokens.

Also extract wait types to session_cmds_daemon_wait_types.rs to stay
under the 8K token monolith limit.
